+++ src/AllanVarianceComputor.cpp
@@ -101,16 +101,16 @@ void AllanVarianceComputor::run(std::string bag_path) {
 void AllanVarianceComputor::closeOutputs() { av_output_.close(); }
 
 void AllanVarianceComputor::allanVariance() {
-  std::vector<std::vector<float>> allan_variances;
+  std::vector<std::vector<double>> allan_variances;
 
   for (int period = 1; period < 10000; period++) {
-    std::vector<std::vector<float>> averages;
-    float period_time = period * 0.1;
+    std::vector<std::vector<double>> averages;
+    double period_time = period * 0.1;
 
     bool new_bin = true;
     int bin_size = 0;
 
-    std::vector<float> current_average = {0.0, 0.0, 0.0, 0.0, 0.0, 0.0};
+    std::vector<double> current_average = {0.0, 0.0, 0.0, 0.0, 0.0, 0.0};
 
     // Compute Averages
     for (const auto& measurement : imuBuffer_) {
@@ -153,7 +153,7 @@ void AllanVarianceComputor::allanVariance() {
 
     // Compute Allan Variance
 
-    std::vector<float> allan_variance = {0.0, 0.0, 0.0, 0.0, 0.0, 0.0};
+    std::vector<double> allan_variance = {0.0, 0.0, 0.0, 0.0, 0.0, 0.0};
     for (int k = 0; k < num_averages - 1; k++) {
       allan_variance[0] += std::pow(averages[k + 1][0] - averages[k][0], 2);
       allan_variance[1] += std::pow(averages[k + 1][1] - averages[k][1], 2);
@@ -162,12 +162,12 @@ void AllanVarianceComputor::allanVariance() {
       allan_variance[4] += std::pow(averages[k + 1][4] - averages[k][4], 2);
       allan_variance[5] += std::pow(averages[k + 1][5] - averages[k][5], 2);
     }
-    std::vector<float> avar = {
+    std::vector<double> avar = {
         allan_variance[0] / (2 * (num_averages - 1)), allan_variance[1] / (2 * (num_averages - 1)),
         allan_variance[2] / (2 * (num_averages - 1)), allan_variance[3] / (2 * (num_averages - 1)),
         allan_variance[4] / (2 * (num_averages - 1)), allan_variance[5] / (2 * (num_averages - 1))};
 
-    std::vector<float> allan_deviation = {std::sqrt(avar[0]), std::sqrt(avar[1]), std::sqrt(avar[2]),
+    std::vector<double> allan_deviation = {std::sqrt(avar[0]), std::sqrt(avar[1]), std::sqrt(avar[2]),
                                           std::sqrt(avar[3]), std::sqrt(avar[4]), std::sqrt(avar[5])};
 
     writeAllanDeviation(allan_deviation, period_time);
@@ -180,7 +180,7 @@ void AllanVarianceComputor::allanVariance() {
   }
 }
 
-void AllanVarianceComputor::writeAllanDeviation(std::vector<float> variance, float period) {
+void AllanVarianceComputor::writeAllanDeviation(std::vector<double> variance, double period) {
   aVRecorder_.period = period;
   aVRecorder_.accX = variance[0];
   aVRecorder_.accY = variance[1];
